requirements.txt
crewai
groq
langchain-groq==0.1.5
langchain-core==0.1.52

flask
speechrecognition
pyttsx3
pyaudio
python-dotenv
websockets
asyncio-mqtt
crewai-tools
litellm
google-generativeai
